테스트중

+++ src/main/java/org/test/teamproject_back/service/AdminService.java
@@ -10,12 +10,16 @@ import org.springframework.transaction.annotation.Transactional;
 import org.test.teamproject_back.dto.request.ReqModifyUserDto;
 import org.test.teamproject_back.dto.response.RespAdminDto;
 import org.test.teamproject_back.entity.User;
+import org.test.teamproject_back.entity.UserRoles;
+import org.test.teamproject_back.exception.InvalidInputException;
 import org.test.teamproject_back.repository.AdminMapper;
 import org.test.teamproject_back.repository.UserMapper;
+import org.test.teamproject_back.repository.UserRolesMapper;
 import org.test.teamproject_back.security.principal.PrincipalUser;
 
 import java.sql.SQLException;
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -26,26 +30,14 @@ public class AdminService {
     @Autowired
     private AdminMapper adminMapper;
     @Autowired
+    private UserRolesMapper userRolesMapper;
+    @Autowired
     private BCryptPasswordEncoder bCryptPasswordEncoder;
 
     public RespAdminDto getAllUsers(String roleName) {
         User user = null;
 
-        PrincipalUser principalUser = (PrincipalUser) SecurityContextHolder
-                .getContext()
-                .getAuthentication()
-                .getPrincipal();
-
-        List<String> authorities = principalUser.getAuthorities().stream()
-                .map(GrantedAuthority::getAuthority)
-                .collect(Collectors.toList());
-
-        if (!(authorities.contains("ROLE_ADMIN") || authorities.contains("ROLE_MANAGER"))) {
-            throw new AuthenticationServiceException("조회 할 수 있는 권한이 없습니다.");
-        }
-
         if ("user".equals(roleName.trim())) {
-            System.out.println("ddd");
             user = userMapper.findAllUsersByRole("ROLE_USER");
         } else if ("manager".equals(roleName.trim())) {
             user = userMapper.findAllUsersByRole("ROLE_MANAGER");
@@ -70,11 +62,14 @@ public class AdminService {
         if (!authorities.contains("ROLE_ADMIN")) {
             throw new AuthenticationServiceException("삭제 할 수 있는 권한이 없습니다.");
         }
-
+        if(!(Optional.ofNullable(adminMapper.findUserById(userId))).isPresent()) {
+            throw new InvalidInputException("해당 사용자 정보가 존재하지 않습니다.");
+        }
         adminMapper.deleteUserByUserId(userId);
-        adminMapper.deleteUserRolesByUserId(userId);
+        userRolesMapper.deleteUserRolesByUserId(userId);
     }
 
+    @Transactional(rollbackFor = SQLException.class)
     public void modifyUser(ReqModifyUserDto dto) {
         PrincipalUser principalUser = (PrincipalUser) SecurityContextHolder
                 .getContext()
@@ -89,7 +84,10 @@ public class AdminService {
             throw new AuthenticationServiceException("수정 할 수 있는 권한이 없습니다.");
         }
 
-        adminMapper.updateByUserId(dto.toEntity(bCryptPasswordEncoder));
+        if(!(Optional.ofNullable(adminMapper.findUserById(dto.getUserId()))).isPresent()) {
+            throw new InvalidInputException("해당 사용자 정보가 존재하지 않습니다.");
+        }
+        adminMapper.updateUserByUserId(dto.toEntity(bCryptPasswordEncoder));
     }
 
 }
